Add files via upload
Using Facebook

+++ posts/views.py
@@ -1,5 +1,5 @@
 from django.contrib import messages
-from django.http import HttpResponse,HttpResponseRedirect
+from django.http import HttpResponse,HttpResponseRedirect,Http404
 from django.shortcuts import render, get_object_or_404,redirect
 from .models import Post
 from .forms import PostForm
@@ -8,19 +8,23 @@ from urllib import quote_plus
 
 # Create your views here.
 def post_detail(request,slug=None):
-    #instance = Post.objects.get(id=3)
+    # instance = Post.objects.get(slug=slug)
     instance = get_object_or_404(Post, slug=slug)
     # building up a query string to go into a URL
     share_string = quote_plus(instance.content)
-
     context = {
         "title":"detail",
          "instance":instance,
-        "share_string":share_string
+        "share_string":share_string,
+
     }
     return render(request, "post_detail.html", context)
 
 def post_delete(request, slug=None):
+    if not request.user.is_staff or not request.user.is_superuser:
+        raise Http404
+    # if not request.user.is_authenticated:
+    #     raise Http404
     instance = get_object_or_404(Post, slug=slug)
     instance.delete()
     messages.success(request, "delete successfully")
@@ -28,14 +32,17 @@ def post_delete(request, slug=None):
     return redirect("posts:list")
 
 def post_create(request):
+    if not request.user.is_staff or not request.user.is_superuser:
+        raise Http404
+    # if not request.user.is_authenticated:
+    #     raise Http404
     form = PostForm(request.POST or None, request.FILES or None)
     if form.is_valid():
         instance = form.save(commit=False)
+        #assume that the user is logged in
+        instance.user = request.user
         instance.save()
         messages.success(request, "Success Created", extra_tags="some_tags")
-    # if request.method == 'POST':
-    #     print request.POST.get("content")
-    #     print request.POST.get("title")
         return HttpResponseRedirect(instance.get_absolute_url())
 
     context = {
@@ -64,6 +71,10 @@ def post_list(request):
     return render(request,"post_list.html",context_data)
 
 def post_update(request, slug=None):
+    if not request.user.is_staff or not request.user.is_superuser:
+        raise Http404
+    # if not request.user.is_authenticated:
+    #     raise Http404
     instance = get_object_or_404(Post, slug=slug)
     form = PostForm(request.POST or None, request.FILES or None, instance=instance)
     if form.is_valid():
